+++ src/Aws/Route53Domains/Resources/route53domains-2014-05-15.php
@@ -2113,4 +2113,18 @@ return array (
             ),
         ),
     ),
+    'iterators' => array(
+        'ListDomains' => array(
+            'limit_key' => 'MaxItems',
+            'input_token' => 'Marker',
+            'output_token' => 'NextPageMarker',
+            'result_key' => 'Domains',
+        ),
+        'ListOperations' => array(
+            'limit_key' => 'MaxItems',
+            'input_token' => 'Marker',
+            'output_token' => 'NextPageMarker',
+            'result_key' => 'Operations',
+        ),
+    ),
 );